requirements.txt
numpy~=1.21.2
torch~=1.9.0
PyYAML~=5.4.1
click~=8.0.1
gym~=0.19.0